+++ dyeti-backend/src/main/java/ki/agh/dyeti/model/PlansRecipes.java
@@ -19,7 +19,7 @@ import lombok.NoArgsConstructor;
 @IdClass(PlansRecipesId.class)
 public class PlansRecipes {
 
-  @Id private Integer planId;
+  @Id private Long planId;
 
-  @Id private Integer recipeId;
+  @Id private Long recipeId;
 }
